trident-acl-agent: address open Copilot review comments on PR 730
- annotations.rs: UpdateRequest::validate() now rejects a stage/finalize
  server URL whose scheme is not http/https (previously only checked
  presence). http is intentionally still allowed (not https-only) since
  the storm E2E harness points server at a plain-http local Nebraska
  stub.
- orchestrator.rs: handle_stage's initial check_for_update() call now
  reports the node's real current version instead of a hardcoded 0.0.0
  (which Nebraska would otherwise offer an update for on every poll,
  forever), and reuses the module's NEBRASKA_MACHINE_ID_SOURCE constant
  instead of an inline IdSource::MachineIdHashed.
- lib.rs: run_omaha_only's one-shot Nebraska check has the same fix,
  reporting the real current version instead of 0.0.0.
- Moved current_active_version() (and its os-release-reading helpers)
  out of annotations.rs into a new version module, since it's unrelated
  to the annotation wire format and is now used by both the
  orchestrator and the omaha-only path.

Copilot-Session: 98412869-e5c8-4f3f-a97c-fe870db70701
